better error handling and small fixes
